rpm  用户处理

+++ python2-seafutil.spec
@@ -57,24 +57,26 @@ done;
 
 
 %pre
-getent group seafile >/dev/null || groupadd -f -g 873 -r seafile
-if ! getent passwd seafile >/dev/null ; then
-    if ! getent passwd 873 >/dev/null ; then
-      useradd -r -u 873 -g seafile -M -s /sbin/nologin -c "Seafile server" seafile
-    else
-      useradd -r -g seafile -M -s /sbin/nologin -c "Seafile server" seafile
+if [ "$1" = "1" ] ; then
+    getent group seafile >/dev/null || groupadd -f -g 873 -r seafile
+    if ! getent passwd seafile >/dev/null ; then
+        if ! getent passwd 873 >/dev/null ; then
+          useradd -r -u 873 -g seafile -M -s /sbin/nologin -c "Seafile server" seafile
+        else
+          useradd -r -g seafile -M -s /sbin/nologin -c "Seafile server" seafile
+        fi
     fi
 fi
-exit 0
 
 
 %preun
 systemctl stop seafile.service
-/usr/sbin/userdel seafile > /dev/null
 
 
-#%postun
-#/usr/sbin/userdel seafile > /dev/null
+%postun
+if [ "$1" = "0" ] ; then
+    /usr/sbin/userdel seafile > /dev/null
+fi
 
 
 %files
